Fix koopa holding turning
Khi hold một hướng thì mario render trước, khi đổi hướng thì mario render sau shell

+++ Camera.h
@@ -31,7 +31,7 @@ class CCamera : public CGameObject
 	//const float levelHeight = 626;
 	float screenWidth = 320;
 	float screenHeight = 240;
-	int shakeStart = -1;
+	ULONGLONG shakeStart = -1;
 public:
 	CMario* GetMario();
